[AIRFLOW-6721] Move HivePartitionSensor tests
Move the HivePartitionSensor tests from the operators module into
the a new test module.
